Improve code quality and code documentation
Adds detailed documentation (courtesy of ChatGPT) to most items. The only thing lacking here is documentation for generics, but it's 11:59 and I am too tired to feel like working on that tonight. Also refactors many extremely long methods into smaller chunks that are a lot more readable and a lot less spagetti.
Closes #13